Created geometry utils calculations + fixed solar 2d vec calculation
